Updated PR comment example syntax.
The comment PR example using steps.preview.outputs.output was not
working despite naming the job 'preview' as defined. While
troubleshooting I found examples that made reference to setting a
specific step id that could then be referenced to get its output. This
worked for me so I updated the documentation to show to use this more
explicit reference.